improvements to Unused Images code
WIP
